Move clean_jedi_cache fixture to top-level conftest.py
Otherwise doctest module running in jedi/ subdirectory will not find it.
